+++ frontend/src/api/schema.d.ts
@@ -2317,8 +2317,6 @@ export interface components {
             allow_remote_images?: boolean | null;
             /** Update Check Enabled */
             update_check_enabled?: boolean | null;
-            /** Network Proxy */
-            network_proxy?: string | null;
             /** Contacts Auto Collect */
             contacts_auto_collect?: boolean | null;
         };